+++ app/library/config.php
@@ -51,7 +51,7 @@ if (!defined('USER_DISPLAYABLE_EXCEPTION'))
     define('USER_DISPLAYABLE_EXCEPTION', 77);
 }
 
-// Since global variables are considered poor form in PHP, these are also
+// Since global variables are considered poor form in PHP, these are
 // defined as constants which could eventually replace the global variables -
 // although a settings table would be the best solution
 
@@ -130,20 +130,33 @@ if (!defined('GOOGLE_RECAPTCHA_V3_ACTION_RESET_PASSWORD'))
     define('GOOGLE_RECAPTCHA_V3_ACTION_RESET_PASSWORD', 'password_reset');
 }
 
+/**
+ * Allow $config variables to be set within Codeception PHPBrowser tests
+ * Webdriver doesn't support the $I->haveHttpHeader method, so we can't set
+ * headers this way.
+ * TODO: figure out a better way of injecting custom qt.ini config values into
+ * tests.
+ */
 if ('Codeception' == ($_SERVER['HTTP_X_REQUESTED_WITH'] ?? ''))
 {
     // Update some config vars from codeception
+    if (isset($_SERVER['HTTP_X_PRIVACY_POLICY_URL']))
+    {
+        $config['privacy_policy_url'] = $_SERVER['HTTP_X_PRIVACY_POLICY_URL'];
+    }
+
+    if (isset($_SERVER['HTTP_X_IS_USER_REGISTRATION_ALLOWED']))
+    {
+        $config['is_user_registration_allowed'] = $_SERVER['HTTP_X_IS_USER_REGISTRATION_ALLOWED'];
+    }
+
     if (isset($_SERVER['HTTP_X_HOSTING_ORGANISATION']))
     {
         $config['hosting_organisation'] = $_SERVER['HTTP_X_HOSTING_ORGANISATION'];
     }
+
     if (isset($_SERVER['HTTP_X_HOSTING_ORGANISATION_URL']))
     {
-        var_dump('HTTP_X_HOSTING_ORGANISATION_URL is set');
         $config['hosting_organisation_url'] = $_SERVER['HTTP_X_HOSTING_ORGANISATION_URL'];
-    }
-    else
-    {
-        var_dump('HTTP_X_HOSTING_ORGANISATION_URL is NOT set');
     }
 } ;