+++ tests/acceptance_phpbrowser/auth/LoginCest.php
@@ -21,6 +21,7 @@ class LoginCest extends QTPageCest
 
     public function _after($I, $scenario)
     {
+        $I->clearLoginLogs($I, ['user_ids' => []]);
     }
 
     public function accessRulesWork(
@@ -39,10 +40,6 @@ class LoginCest extends QTPageCest
     public function showsNoHostingOrganisationIfNotSet(AcceptancePhpbrowserTester $I)
     {
         global $config;
-        codecept_debug($config);
-
-        $codecept_config        = [];
-        $codecept_config['foo'] = 'bar';
 
         // this is a hack. Sending a space sets the header, but ultimately
         // resets the configs back to nothing by the time we get to login.php
@@ -63,10 +60,6 @@ class LoginCest extends QTPageCest
     public function showsHostingOrganisationAndUrlIfSet(AcceptancePhpbrowserTester $I)
     {
         global $config;
-        codecept_debug($config);
-
-        $codecept_config        = [];
-        $codecept_config['foo'] = 'bar';
 
         $hosting_organisation     = 'University X';
         $hosting_organisation_url = 'https://universityx.edu/';
@@ -87,10 +80,6 @@ class LoginCest extends QTPageCest
     public function showsOnlyHostingOrganisationIfUrlNotSet(AcceptancePhpbrowserTester $I)
     {
         global $config;
-        codecept_debug($config);
-
-        $codecept_config        = [];
-        $codecept_config['foo'] = 'bar';
 
         $hosting_organisation = 'University Y';
         // this is a hack. Sending a space sets the header, but ultimately